* commit 'b26e4944e2a5bf4b7e33fa38fb17a3b76c7fe0f7':
  Fix 64-bit build.

+++ libc/bionic/system_properties.cpp
@@ -238,7 +238,7 @@ static int map_fd_ro(const int fd) {
     if ((fd_stat.st_uid != 0)
             || (fd_stat.st_gid != 0)
             || ((fd_stat.st_mode & (S_IWGRP | S_IWOTH)) != 0)
-            || (fd_stat.st_size < sizeof(prop_area)) ) {
+            || (fd_stat.st_size < static_cast<off_t>(sizeof(prop_area))) ) {
         return -1;
     }
